Extract base notifiable block and tile
