+++ src/main/scala/ru/mousecray/endmagic/blocks/portal/BlockMasterBasePortal.java
@@ -1,6 +1,5 @@
 package ru.mousecray.endmagic.blocks.portal;
 
-import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
@@ -8,31 +7,17 @@ import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
-import net.minecraft.world.World;
-import ru.mousecray.endmagic.blocks.BlockWithTile;
+import ru.mousecray.endmagic.blocks.BlockWithNotifiableTile;
 import ru.mousecray.endmagic.tileentity.portal.TileMasterBasePortal;
 
 import java.util.function.Supplier;
 
-public class BlockMasterBasePortal<A extends TileMasterBasePortal> extends BlockWithTile<A> {
+public class BlockMasterBasePortal<A extends TileMasterBasePortal> extends BlockWithNotifiableTile<A> {
     public BlockMasterBasePortal(Supplier<A> factory) {
         super(Material.ROCK, factory);
         setResistance(8.0F);
         setHardness(4.0F);
         setSoundType(SoundType.STONE);
-    }
-
-    @Override
-    public void neighborChanged(IBlockState state, World worldIn, BlockPos pos, Block blockIn, BlockPos fromPos) {
-        if (!worldIn.isRemote)
-            tile(worldIn, pos).neighborChanged();
-    }
-
-    @Override
-    public void breakBlock(World worldIn, BlockPos pos, IBlockState state) {
-        if (!worldIn.isRemote)
-            tile(worldIn, pos).breakBlock();
-        super.breakBlock(worldIn, pos, state);
     }
 
     @Override
