+++ src/main/scala/ru/mousecray/endmagic/tileentity/ByBlockNotifiable.java
@@ -1,0 +1,6 @@
+package ru.mousecray.endmagic.tileentity;
+
+public interface ByBlockNotifiable {
+    public void neighborChanged();
+    public void breakBlock();
+}
